+++ man-roxygen/args-sparse.R
@@ -1,0 +1,3 @@
+#' @param sparse A single string. Should the columns produced be sparse vectors.
+#'   Can take the values `"yes"`, `"no"`, and `"auto"`. If `sparse = "auto"`
+#'   then workflows can determine the best option. Defaults to `"auto"`.